BetterController Post-demo Refactor

+++ server/src/main/java/com/ncc/neon/models/BetterFile.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class BetterFile {
     private final String filename;
     // TODO: Set to long type.
-    private final int bytes;
+    private final long bytes;
 }
